Add an interface for Query
